backend/requirements.txt
fastpi
"uvicorn[standard]"
python-decouple
beanie
email-validator
python-jose[cryptography]
"passlib[bcrypt]"
python-multipart